modules/grad_cam_trainer/requirements.txt
certifi==2022.6.15
charset-normalizer==2.0.12
cycler==0.11.0
fonttools==4.33.3
idna==3.3
kiwisolver==1.4.3
matplotlib==3.5.2
numpy==1.22.4
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.2
Pillow==9.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.0
six==1.16.0
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
tqdm==4.64.0
typing-extensions==4.2.0
urllib3==1.26.9
